api/requirements.txt
fastapi
uvicorn
sentence-transformers
qdrant-client
kafka-python
python-multipart